test: split paddle webhook handler tests
