+++ scripts/check-db-access-guard.mjs
@@ -22,6 +22,7 @@ const APPROVED_WRAPPER_PATTERNS = [/^packages\/database\//u];
 const EXCLUDED_PATH_PATTERNS = [
   /(^|\/)__tests__(\/|$)/u,
   /(^|\/)(node_modules|dist|build|coverage|test-results|\.next|\.turbo)(\/|$)/u,
+  /(^|\/)test-support\.[cm]?[jt]sx?$/u,
   /\.(test|spec)\.[cm]?[jt]sx?$/u,
   /\.d\.ts$/u,
 ];
